+++ src/compas_model/algorithms/bvh.py
@@ -7,13 +7,14 @@ from compas.datastructures import Mesh
 from compas.datastructures import Tree
 from compas.datastructures import TreeNode
 from compas.geometry import Box
+from compas.geometry import Frame
 from compas.geometry import Line
 from compas.geometry import Point
 from compas.geometry import centroid_points
-from compas.geometry import oriented_bounding_box_numpy
-from compas_model.algorithms.intersections import is_intersection_box_box
-from compas_model.algorithms.intersections import is_intersection_line_aabb
-from compas_model.algorithms.intersections import is_intersection_line_box
+from compas_model.algorithms import is_intersection_box_box
+from compas_model.algorithms import is_intersection_line_aabb
+from compas_model.algorithms import is_intersection_line_box
+from compas_model.algorithms import pca_box
 
 
 class BVHNode(TreeNode):
@@ -35,6 +36,7 @@ class BVHNode(TreeNode):
     def __init__(self, objects: list[tuple[int, Point, list[Point]]], **kwargs):
         super().__init__(**kwargs)
         self.objects = objects
+        self.depth = 0
         self._box = None
 
     @property
@@ -83,7 +85,7 @@ class OBBNode(BVHNode):
     def compute_box(self) -> Box:
         """"""
         points = [point for o in self.objects for point in o[2]]
-        return Box.from_bounding_box(oriented_bounding_box_numpy(points))
+        return pca_box(points)
 
     def intersect_line(self, line: Line) -> Generator["OBBNode", None, None]:
         queue = [self]
@@ -103,23 +105,69 @@ class OBBNode(BVHNode):
 
 
 class BVH(Tree):
+    """Bounding Volume Hierarchy as a special case of a (binary) tree.
+
+    Parameters
+    ----------
+    nodetype : Type[:class:`AABBNode`] | Type[:class:`OBBNode`], optional
+        The type of boundng volume node to use in the tree.
+    max_depth : int, optional
+        The maximum depth of the tree.
+    leafsize : int, optional
+        The number of objects contained by a leaf.
+
+    Notes
+    -----
+    This class has the following constructors:
+
+    * :meth:`BVH.from_triangles`
+    * :meth:`BVH.from_mesh`
+
+    References
+    ----------
+    ...
+
+    Examples
+    --------
+    >>>
+
+    """
+
     root: BVHNode
+
+    def __init__(
+        self,
+        nodetype: Optional[Union[Type[AABBNode], Type[OBBNode]]] = AABBNode,
+        max_depth: Optional[int] = None,
+        leafsize: int = 1,
+        **kwargs,
+    ):
+        super().__init__(**kwargs)
+        self.nodetype = nodetype
+        self.max_depth = max_depth
+        self.leafsize = leafsize
 
     def _add_objects(
         self,
         objects: list[tuple[int, Point, list[Point]]],
         parent: BVHNode,
-        nodetype: Optional[Union[Type[AABBNode], Type[OBBNode]]] = OBBNode,
     ) -> None:
         if not objects:
             return
 
-        if len(objects) == 1:
-            parent.add(nodetype(objects[:1]))
+        node = self.nodetype(objects)
+        parent.add(node)
+
+        if node.is_root:
+            node.depth = 0
+        else:
+            node.depth = parent.depth + 1
+
+        if len(objects) <= self.leafsize:
             return
 
-        node = nodetype(objects)
-        parent.add(node)
+        if self.max_depth and node.depth >= self.max_depth:
+            return
 
         # sort objects according to xaxis box
         # (use the xaxis because it has the largest spread/variance)
@@ -135,32 +183,38 @@ class BVH(Tree):
         # with left/right instead of a list of children
 
         # "left" objects
-        self._add_objects(objects[:median], parent=node, nodetype=nodetype)
+        self._add_objects(objects[:median], parent=node)
 
         # "right" objects
-        self._add_objects(objects[median:], parent=node, nodetype=nodetype)
+        self._add_objects(objects[median:], parent=node)
 
     @classmethod
     def from_triangles(
         cls,
         triangles: list[list[Point]],
-        nodetype: Optional[Union[Type[AABBNode], Type[OBBNode]]] = OBBNode,
+        nodetype: Optional[Union[Type[AABBNode], Type[OBBNode]]] = AABBNode,
+        max_depth: Optional[int] = None,
+        leafsize: int = 1,
     ) -> "BVH":
         objects = [(index, centroid_points(abc), abc) for index, abc in enumerate(triangles)]
-        tree = cls()
-        tree._add_objects(objects, parent=tree, nodetype=nodetype)
+
+        tree = cls(nodetype=nodetype, max_depth=max_depth, leafsize=leafsize)
+        tree._add_objects(objects, parent=tree)
         return tree
 
     @classmethod
     def from_mesh(
         cls,
         mesh: Mesh,
-        nodetype: Optional[Union[Type[AABBNode], Type[OBBNode]]] = OBBNode,
+        nodetype: Optional[Union[Type[AABBNode], Type[OBBNode]]] = AABBNode,
+        max_depth: Optional[int] = None,
+        leafsize: int = 1,
     ) -> "BVH":
         faces = list(mesh.faces())
         objects = list(zip(faces, [mesh.face_centroid(face) for face in faces], [mesh.face_points(face) for face in faces]))
-        tree = cls()
-        tree._add_objects(objects, parent=tree, nodetype=nodetype)
+
+        tree = cls(nodetype=nodetype, max_depth=max_depth, leafsize=leafsize)
+        tree._add_objects(objects, parent=tree)
         return tree
 
     def intersect_line(self, line: Line) -> Generator[BVHNode, None, None]: